requirements.txt
# --- Core AI & Orchestration ---
langchain>=0.1.0
openai>=1.0.0          
python-dotenv  
mistralai         # Environment variable management (API Keys)

# --- Data Extraction (Text & Images) ---
PyMuPDF>=1.23.0         # Fast PDF text and image extraction 
unstructured[pdf]       # Handles complex document partitioning and tables [cite: 25, 36]
pillow                  # Image processing and resizing for the final DDR [cite: 52]

# --- Report Generation ---
python-docx             # To generate the final structured Word/Client-ready report [cite: 40, 53]
fpdf2                   # Alternative for generating structured PDF reports [cite: 53]
pandas                  # For data structuring and deduplication [cite: 38]

# --- Utilities ---
requests                # For any external API calls or image handling
pydantic                